backend/requirements.txt
# Backend Python dependencies (minimal set). Adjust versions as needed.
Django==5.2.8
djangorestframework
django-cors-headers
django-filter
python-dotenv
requests
gunicorn
psycopg2-binary
Pillow
whitenoise
dj-database-url
